+++ paddle/fluid/jit/CMakeLists.txt
@@ -35,7 +35,7 @@ cc_library(
   jit_function
   SRCS function.cc
   DEPS jit_function_utils jit_executor_engine jit_pe_engine
-       jit_interpreter_engine)
+       jit_interpreter_engine jit_predictor_engine)
 
 cc_library(
   jit_layer
