+++ paddle/fluid/jit/engine/CMakeLists.txt
@@ -12,3 +12,8 @@ cc_library(
   jit_interpreter_engine
   SRCS interpreter_engine.cc
   DEPS standalone_executor)
+
+cc_library(
+  jit_predictor_engine
+  SRCS predictor_engine.cc
+  DEPS paddle_inference_api analysis_predictor)
