+++ paddle/fluid/jit/function_schema.h
@@ -72,6 +72,10 @@ class FunctionInfo {
 
   const std::vector<std::string> OutputArgNames() const;
 
+  const std::string& ProgramFilePath() const;
+
+  void SetProgramFilePath(const std::string& path);
+
   void RemoveDescFeedFetch();
 
  private:
@@ -79,6 +83,7 @@ class FunctionInfo {
   std::vector<std::string> param_names_;
   std::shared_ptr<framework::ProgramDesc> program_desc_;
   FunctionSchema schema_;
+  std::string prog_file_path_;
 };
 
 }  // namespace jit
