+++ paddle/fluid/jit/serializer.cc
@@ -23,6 +23,7 @@
 #include "paddle/fluid/jit/engine/executor_engine.h"
 #include "paddle/fluid/jit/engine/interpreter_engine.h"
 #include "paddle/fluid/jit/engine/pe_engine.h"
+#include "paddle/fluid/jit/engine/predictor_engine.h"
 #include "paddle/fluid/jit/layer.h"
 #include "paddle/fluid/jit/property.h"
 #include "paddle/fluid/jit/serializer_utils.h"
@@ -54,6 +55,7 @@ Layer Deserializer::operator()(const std::string& path,
     param_names_set.insert(persist_var_names.begin(), persist_var_names.end());
     info_map[func_name] = std::make_shared<FunctionInfo>(
         func_name, persist_var_names, program_desc);
+    info_map[func_name]->SetProgramFilePath(it.second);
   }
 
   VariableMap params_dict;
@@ -70,22 +72,23 @@ Layer Deserializer::operator()(const std::string& path,
   for (auto it = info_map.begin(); it != info_map.end(); ++it) {
     const std::string& func_name = it->first;
     auto& info = it->second;
+    VLOG(3) << "Add function type: " << FLAGS_jit_engine_type
+            << " Function name: " << func_name;
     if (FLAGS_jit_engine_type == "Executor") {
-      VLOG(3) << "Add function type: ExecutorEngine. Function name: "
-              << func_name;
       layer.SetEngine(
           func_name,
           utils::MakeEngine<ExecutorEngine>(info, params_dict, place));
     } else if (FLAGS_jit_engine_type == "PE") {
-      VLOG(3) << "Add function type: PEEngine. Function name: " << func_name;
       layer.SetEngine(func_name,
                       utils::MakeEngine<PEEngine>(info, params_dict, place));
     } else if (FLAGS_jit_engine_type == "New") {
-      VLOG(3) << "Add function type: InterpreterEngine. Function name: "
-              << func_name;
       layer.SetEngine(
           func_name,
           utils::MakeEngine<InterpreterEngine>(info, params_dict, place));
+    } else if (FLAGS_jit_engine_type == "Predictor") {
+      layer.SetEngine(
+          info->FunctionName(),
+          utils::MakeEngine<PredictorEngine>(info, params_dict, place));
     } else {
       PD_THROW("Invalid JitLayer engine type.");
     }
